# app/src/main/java/br/com/nsfatima/calendario/api/dto/evento/EventoResponse.java
package br.com.nsfatima.calendario.api.dto.evento;

import java.time.Instant;
import java.util.UUID;

public record EventoResponse(
        UUID id,
        String titulo,
        String descricao,
        Instant inicio,
        Instant fim,
        String status) {
}