# specs/001-parish-calendar-api/contracts/calendar-api.openapi.yaml
openapi: 3.0.3
info:
  title: Calendario Paroquial API
  version: 0.1.0
  description: API para gerenciamento de eventos, projetos e observacoes do calendario paroquial.
  x-implementation-status: phase-6-polish
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /eventos:
    post:
      summary: Criar evento paroquial
      operationId: createEvento
      tags: [Eventos]
      description: Requer autenticacao e RBAC por organizacao. Se `status=ADICIONADO_EXTRA`, justificativa obrigatoria deve ser enviada.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventoRequest'
      responses:
        '201':
          description: Evento criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    get:
      summary: Listar eventos por periodo
      operationId: listEventos
      tags: [Eventos]
      description: Consulta de calendario publico; para acesso anonimo retorna somente eventos CONFIRMADO. Eventos RASCUNHO nunca aparecem neste endpoint publico.
      security: []
      parameters:
        - in: query
          name: inicio
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: fim
          required: true
          schema: { type: string, format: date-time }
        - in: query
          name: projetoId
          required: false
          schema: { type: string, format: uuid }
        - in: query
          name: categoriaId
          required: false
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Lista de eventos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

  /eventos/{eventoId}:
    get:
      summary: Obter evento por id
      operationId: getEventoById
      tags: [Eventos]
      description: Leitura interna autenticada; aplica RBAC hierarquico e escopo por organizacao.
      parameters:
        - $ref: '#/components/parameters/EventoId'
      responses:
        '200':
          description: Evento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      summary: Atualizar evento
      operationId: patchEvento
      tags: [Eventos]
      parameters:
        - $ref: '#/components/parameters/EventoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventoRequest'
      responses:
        '200':
          description: Evento atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Cancelar evento (soft delete)
      operationId: cancelEvento
      tags: [Eventos]
      parameters:
        - $ref: '#/components/parameters/EventoId'
      responses:
        '204':
          description: Evento cancelado
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /eventos/{eventoId}/participantes:
    put:
      summary: Definir organizacoes participantes de um evento
      operationId: putEventoParticipantes
      tags: [Eventos]
      description: Operacao permitida apenas para Admin Geral e Admin de Pastoral da organizacao responsavel do evento.
      parameters:
        - $ref: '#/components/parameters/EventoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventoParticipantesRequest'
      responses:
        '200':
          description: Participantes atualizados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Limpar todas as organizacoes participantes de um evento
      operationId: clearEventoParticipantes
      tags: [Eventos]
      parameters:
        - $ref: '#/components/parameters/EventoId'
      responses:
        '200':
          description: Participantes removidos; evento permanece valido sem participantes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /projetos:
    post:
      summary: Criar projeto pastoral
      operationId: createProjeto
      tags: [Projetos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjetoRequest'
      responses:
        '201':
          description: Projeto criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjetoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    get:
      summary: Listar projetos
      operationId: listProjetos
      tags: [Projetos]
      responses:
        '200':
          description: Lista de projetos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjetoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /projetos/{projetoId}:
    patch:
      summary: Atualizar projeto
      operationId: patchProjeto
      tags: [Projetos]
      parameters:
        - $ref: '#/components/parameters/ProjetoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjetoRequest'
      responses:
        '200':
          description: Projeto atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjetoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /eventos/{eventoId}/observacoes:
    post:
      summary: Incluir observacao no evento
      operationId: createObservacao
      tags: [Observacoes]
      description: Observacao e append-only e pode ser utilizada para rastrear justificativas obrigatorias de eventos ADICIONADO_EXTRA.
      parameters:
        - $ref: '#/components/parameters/EventoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateObservacaoRequest'
      responses:
        '201':
          description: Observacao criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservacaoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    get:
      summary: Listar observacoes de um evento
      operationId: listObservacoesByEvento
      tags: [Observacoes]
      parameters:
        - $ref: '#/components/parameters/EventoId'
      responses:
        '200':
          description: Lista de observacoes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObservacaoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /aprovacoes:
    post:
      summary: Submeter solicitacao sensivel (horario, cancelamento ou remocao administrativa)
      operationId: createSolicitacaoAprovacao
      tags: [Aprovacoes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAprovacaoRequest'
      responses:
        '201':
          description: Solicitacao registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AprovacaoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /aprovacoes/{aprovacaoId}/decisao:
    post:
      summary: Aprovar ou reprovar solicitacao
      operationId: decideAprovacao
      tags: [Aprovacoes]
      parameters:
        - $ref: '#/components/parameters/AprovacaoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisaoAprovacaoRequest'
      responses:
        '200':
          description: Decisao registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AprovacaoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    EventoId:
      in: path
      name: eventoId
      required: true
      schema: { type: string, format: uuid }
    ProjetoId:
      in: path
      name: projetoId
      required: true
      schema: { type: string, format: uuid }
    AprovacaoId:
      in: path
      name: aprovacaoId
      required: true
      schema: { type: string, format: uuid }

  responses:
    ValidationError:
      description: Erro de validacao
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Acesso negado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Autenticacao obrigatoria ou token invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    CreateEventoRequest:
      type: object
      required:
        - titulo
        - tipoAtividade
        - categoriaId
        - localId
        - organizacaoResponsavelId
        - inicio
        - fim
        - status
      properties:
        titulo: { type: string, minLength: 1, maxLength: 160 }
        descricao: { type: string, maxLength: 4000 }
        tipoAtividade: { type: string }
        categoriaId: { type: string, format: uuid }
        localId: { type: string, format: uuid }
        projetoId: { type: string, format: uuid, nullable: true }
        organizacaoResponsavelId: { type: string, format: uuid }
        inicio: { type: string, format: date-time, description: UTC }
        fim: { type: string, format: date-time, description: UTC }
        status:
          type: string
          enum: [RASCUNHO, CONFIRMADO, CANCELADO, ADICIONADO_EXTRA]
        adicionadoExtraJustificativa:
          type: string
          minLength: 1
          maxLength: 4000
          description: Obrigatoria quando status=ADICIONADO_EXTRA.
      example:
        titulo: Missa Dominical
        descricao: Celebracao da comunidade
        tipoAtividade: MISSA
        categoriaId: 11111111-1111-1111-1111-111111111111
        localId: 22222222-2222-2222-2222-222222222222
        organizacaoResponsavelId: 33333333-3333-3333-3333-333333333333
        inicio: "2026-03-15T10:00:00Z"
        fim: "2026-03-15T11:00:00Z"
        status: CONFIRMADO

    UpdateEventoRequest:
      type: object
      properties:
        titulo: { type: string, minLength: 1, maxLength: 160 }
        descricao: { type: string, maxLength: 4000 }
        categoriaId: { type: string, format: uuid }
        localId: { type: string, format: uuid }
        inicio: { type: string, format: date-time, description: UTC }
        fim: { type: string, format: date-time, description: UTC }
        status:
          type: string
          enum: [RASCUNHO, CONFIRMADO, CANCELADO, ADICIONADO_EXTRA]
        adicionadoExtraJustificativa:
          type: string
          minLength: 1
          maxLength: 4000
          description: Obrigatoria quando status=ADICIONADO_EXTRA.

    EventoParticipantesRequest:
      type: object
      required: [organizacoesParticipantes]
      properties:
        organizacoesParticipantes:
          type: array
          minItems: 1
          items: { type: string, format: uuid }

    EventoResponse:
      type: object
      required:
        - id
        - titulo
        - organizacaoResponsavelId
        - organizacoesParticipantes
        - inicio
        - fim
        - status
      properties:
        id: { type: string, format: uuid }
        titulo: { type: string }
        descricao: { type: string }
        projetoId: { type: string, format: uuid, nullable: true }
        organizacaoResponsavelId: { type: string, format: uuid }
        organizacoesParticipantes:
          type: array
          items: { type: string, format: uuid }
        inicio:
          type: string
          format: date-time
          description: America/Sao_Paulo
        fim:
          type: string
          format: date-time
          description: America/Sao_Paulo
        status:
          type: string
          enum: [RASCUNHO, CONFIRMADO, CANCELADO, ADICIONADO_EXTRA]
        temConflito: { type: boolean }

    CreateProjetoRequest:
      type: object
      required: [nome]
      properties:
        nome: { type: string, minLength: 1, maxLength: 160 }
        descricao: { type: string, maxLength: 2000 }

    UpdateProjetoRequest:
      type: object
      properties:
        nome: { type: string, minLength: 1, maxLength: 160 }
        descricao: { type: string, maxLength: 2000 }

    ProjetoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        nome: { type: string }
        descricao: { type: string }
        status: { type: string }

    CreateObservacaoRequest:
      type: object
      required: [usuarioId, tipo, conteudo]
      properties:
        usuarioId: { type: string, format: uuid }
        tipo: { type: string }
        conteudo: { type: string, minLength: 1, maxLength: 4000 }

    ObservacaoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        eventoId: { type: string, format: uuid }
        usuarioId: { type: string, format: uuid }
        tipo: { type: string }
        conteudo: { type: string }
        criadoEm: { type: string, format: date-time }

    CreateAprovacaoRequest:
      type: object
      required: [eventoId, tipoSolicitacao, solicitanteUsuarioId, justificativa]
      properties:
        eventoId: { type: string, format: uuid }
        tipoSolicitacao:
          type: string
          enum: [ALTERACAO_HORARIO, CANCELAMENTO, REMOCAO_ADMINISTRATIVA]
        solicitanteUsuarioId: { type: string, format: uuid }
        justificativa: { type: string, minLength: 1, maxLength: 2000 }

    DecisaoAprovacaoRequest:
      type: object
      required: [decisorUsuarioId, decisao, justificativa]
      properties:
        decisorUsuarioId: { type: string, format: uuid }
        decisao: { type: string, enum: [APROVADO, REPROVADO] }
        justificativa: { type: string, minLength: 1, maxLength: 2000 }

    AprovacaoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        eventoId: { type: string, format: uuid }
        tipoSolicitacao: { type: string }
        status: { type: string }
        solicitanteUsuarioId: { type: string, format: uuid }
        decisorUsuarioId: { type: string, format: uuid, nullable: true }
        criadoEm: { type: string, format: date-time }
        decididoEm: { type: string, format: date-time, nullable: true }

    ErrorResponse:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, AUTH_REQUIRED, ACCESS_DENIED, RESOURCE_NOT_FOUND, CALENDAR_CONFLICT, DOMAIN_RULE_VIOLATION]
        message: { type: string }
        correlationId: { type: string }
      example:
        code: ACCESS_DENIED
        message: Usuario sem permissao para a operacao
        correlationId: cec6f0a8-2fec-4620-a42a-49b58966189d